requirements.txt
aiofiles==23.2.1
aiohttp==3.9.3
aiosignal==1.3.1
altair==5.2.0
annotated-types==0.6.0
anyio==4.2.0
appdirs==1.4.4
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
arrow==1.3.0
asttokens==2.4.1
async-lru==2.0.4
attrs==23.2.0
Authlib==1.3.0
Babel==2.14.0
beautifulsoup4==4.12.3
bleach==6.1.0
bs4==0.0.2
cachetools==5.3.2
certifi==2024.2.2
cffi==1.16.0
chardet==5.2.0
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
comm==0.2.1
contourpy==1.2.0
cryptography==42.0.3
cycler==0.12.1
dataclasses-json==0.6.4
debugpy==1.8.1
decorator==5.1.1
defusedxml==0.7.1
Deprecated==1.2.14
dirtyjson==1.0.8
discord-webhook==1.3.1
distro==1.9.0
docker-pycreds==0.4.0
executing==2.0.1
fastapi==0.109.2
fastjsonschema==2.19.1
ffmpy==0.3.2
filelock==3.13.1
fonttools==4.49.0
fqdn==1.5.1
frozenlist==1.4.1
fsspec==2024.2.0
gitdb==4.0.11
GitPython==3.1.42
google-ai-generativelanguage==0.4.0
google-api-core==2.17.0
google-auth==2.27.0
google-generativeai==0.3.2
googleapis-common-protos==1.62.0
gradio==4.19.1
gradio_client==0.10.0
greenlet==3.0.3
grpcio==1.60.1
grpcio-status==1.60.1
h11==0.14.0
html2text==2020.1.16
httpcore==1.0.2
httpx==0.26.0
huggingface-hub==0.20.3
idna==3.6
importlib-resources==6.1.1
ipykernel==6.29.2
ipython==8.21.0
ipywidgets==8.1.2
isoduration==20.11.0
jedi==0.19.1
Jinja2==3.1.3
joblib==1.3.2
json5==0.9.14
jsonpointer==2.4
jsonschema==4.21.1
jsonschema-specifications==2023.12.1
jupyter-events==0.9.0
jupyter-lsp==2.2.2
jupyter_client==8.6.0
jupyter_core==5.7.1
jupyter_server==2.12.5
jupyter_server_terminals==0.5.2
jupyterlab==4.1.1
jupyterlab_pygments==0.3.0
jupyterlab_server==2.25.2
jupyterlab_widgets==3.0.10
kiwisolver==1.4.5
llama-hub==0.0.79.post1
llama-index==0.10.3
llama-index-agent-openai==0.1.1
llama-index-callbacks-wandb==0.1.1
llama-index-core==0.10.3
llama-index-embeddings-gemini==0.1.1
llama-index-embeddings-openai==0.1.1
llama-index-legacy==0.9.48
llama-index-llms-gemini==0.1.1
llama-index-llms-openai==0.1.1
llama-index-multi-modal-llms-openai==0.1.1
llama-index-program-openai==0.1.1
llama-index-question-gen-openai==0.1.1
llama-index-readers-file==0.1.3
llama-index-retrievers-bm25==0.1.1
llama-index-vector-stores-weaviate==0.1.1
llmsherpa==0.1.4
lxml==5.1.0
markdown-it-py==3.0.0
MarkupSafe==2.1.5
marshmallow==3.20.2
matplotlib==3.8.3
matplotlib-inline==0.1.6
mdurl==0.1.2
mistune==3.0.2
multidict==6.0.5
mypy-extensions==1.0.0
nbclient==0.9.0
nbconvert==7.16.0
nbformat==5.9.2
nest-asyncio==1.6.0
networkx==3.2.1
nltk==3.8.1
notebook_shim==0.2.3
numpy==1.26.4
openai==1.12.0
orjson==3.9.14
overrides==7.7.0
packaging==23.2
pandas==2.2.0
pandocfilters==1.5.1
parso==0.8.3
pexpect==4.9.0
pillow==10.2.0
platformdirs==4.2.0
prometheus-client==0.19.0
prompt-toolkit==3.0.43
proto-plus==1.23.0
protobuf==4.25.2
psutil==5.9.8
ptyprocess==0.7.0
pure-eval==0.2.2
pyaml==23.12.0
pyarrow==15.0.0
pyasn1==0.5.1
pyasn1-modules==0.3.0
pycparser==2.21
pydantic==2.6.1
pydantic_core==2.16.2
pydub==0.25.1
Pygments==2.17.2
PyMuPDF==1.23.22
PyMuPDFb==1.23.22
pyparsing==3.1.1
pypdf==4.0.1
python-dateutil==2.8.2
python-dotenv==1.0.1
python-json-logger==2.0.7
python-multipart==0.0.9
pytz==2024.1
PyYAML==6.0.1
pyzmq==25.1.2
rank-bm25==0.2.2
referencing==0.33.0
regex==2023.12.25
requests==2.31.0
retrying==1.3.4
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rich==13.7.0
rpds-py==0.18.0
rsa==4.9
ruff==0.2.1
semantic-version==2.10.0
Send2Trash==1.8.2
sentry-sdk==1.40.4
setproctitle==1.3.3
shellingham==1.5.4
six==1.16.0
smmap==5.0.1
sniffio==1.3.0
soupsieve==2.5
SQLAlchemy==2.0.27
stack-data==0.6.3
starlette==0.36.3
tabulate==0.9.0
tenacity==8.2.3
terminado==0.18.0
tiktoken==0.6.0
tinycss2==1.2.1
tomlkit==0.12.0
toolz==0.12.1
tornado==6.4
tqdm==4.66.2
traitlets==5.14.1
typer==0.9.0
types-python-dateutil==2.8.19.20240106
typing-inspect==0.9.0
typing_extensions==4.9.0
tzdata==2024.1
uri-template==1.3.0
urllib3==2.2.0
uvicorn==0.27.1
validators==0.22.0
wandb==0.16.3
wcwidth==0.2.13
weaviate-client==3.26.2
webcolors==1.13
webencodings==0.5.1
websocket-client==1.7.0
websockets==11.0.3
widgetsnbextension==4.0.10
wrapt==1.16.0
yarl==1.9.4
